+++ superset-frontend/cypress-base/cypress/integration/dashboard/_skip.controls.test.ts
@@ -25,14 +25,14 @@ import { WORLD_HEALTH_CHARTS } from './utils';
 import { WORLD_HEALTH_DASHBOARD } from 'cypress/utils/urls';
 import { isLegacyResponse } from '../../utils/vizPlugins';
 
-describe('Dashboard top-level controls', () => {
+describe.skip('Dashboard top-level controls', () => {
   beforeEach(() => {
     cy.login();
     cy.visit(WORLD_HEALTH_DASHBOARD);
   });
 
   // flaky test
-  xit('should allow chart level refresh', () => {
+  it('should allow chart level refresh', () => {
     const mapSpec = WORLD_HEALTH_CHARTS.find(
       ({ viz }) => viz === 'world_map',
     ) as ChartSpec;
@@ -58,7 +58,7 @@ describe('Dashboard top-level controls', () => {
     });
   });
 
-  xit('should allow dashboard level force refresh', () => {
+  it('should allow dashboard level force refresh', () => {
     // when charts are not start loading, for example, under a secondary tab,
     // should allow force refresh
     WORLD_HEALTH_CHARTS.forEach(waitForChartLoad);
