+++ superset-frontend/cypress-base/cypress/integration/dashboard/_skip.filter.test.ts
@@ -24,13 +24,13 @@ import {
 import {WORLD_HEALTH_CHARTS } from './utils';
 import { WORLD_HEALTH_DASHBOARD } from 'cypress/utils/urls';
 
-describe('Dashboard filter', () => {
+describe.skip('Dashboard filter', () => {
   before(() => {
     cy.login();
     cy.visit(WORLD_HEALTH_DASHBOARD);
   });
 
-  xit('should apply filter', () => {
+  it('should apply filter', () => {
     WORLD_HEALTH_CHARTS.forEach(waitForChartLoad);
     getChartAliasesBySpec(
       WORLD_HEALTH_CHARTS.filter(({ viz }) => viz !== 'filter_box'),
