+++ superset-frontend/cypress-base/cypress/integration/dashboard/_skip.url_params.test.ts
@@ -24,7 +24,7 @@ import {
 import { WORLD_HEALTH_CHARTS } from './utils';
 import { WORLD_HEALTH_DASHBOARD } from 'cypress/utils/urls';
 
-describe('Dashboard form data', () => {
+describe.skip('Dashboard form data', () => {
   const urlParams = { param1: '123', param2: 'abc' };
   before(() => {
     cy.login();
@@ -32,7 +32,7 @@ describe('Dashboard form data', () => {
     cy.visit(WORLD_HEALTH_DASHBOARD, { qs: urlParams });
   });
 
-  xit('should apply url params to slice requests', () => {
+  it('should apply url params to slice requests', () => {
     cy.intercept('/api/v1/chart/data?*', request => {
       // TODO: export url params to chart data API
       request.body.queries.forEach((query: { url_params: JsonObject }) => {
