+++ superset-frontend/cypress-base/cypress/utils/urls.ts
@@ -21,3 +21,4 @@ export const DASHBOARD_LIST = '/dashboard/list/';
 export const CHART_LIST = '/chart/list/';
 export const WORLD_HEALTH_DASHBOARD = '/superset/dashboard/world_health/';
 export const SAMPLE_DASHBOARD_1 = '/superset/dashboard/1-sample-dashboard/';
+export const ECHARTS_DASHBOARD = '/superset/dashboard/echarts_dash/';
